affichage d'une note première versions, pas encore d'éd
