+++ OL13/aff_notes.cpp
@@ -6,14 +6,28 @@ page_notes::page_notes(Note& n)
     layout_p=new QVBoxLayout(this);
     layout_titre=new QHBoxLayout;
     titre=new QLabel(n.getTitle());
-    titre->setEnabled(false);
     info=new QLabel("Cette note est un :"+n.getType()+" Créer le :"+n.getCreation_date().toString()+" Dernière modification le :"+n.getLastmodif_date().toString());
     layout_titre->addStretch();
     layout_titre->addWidget(titre);
     layout_titre->addStretch();
     layout_p->addLayout(layout_titre);
+    if(n.getType()=="Recording")
+    {
+        note=new QRecording;
 
+    }
+    else if(n.getType()=="Task")
+    {
+        note=new QTask;
+    }
+    else{
+        note=new QArticle;
+    }
+    note->load_note(n);
 
+    layout_p->addWidget(note);
+    layout_p->addStretch();
+    layout_p->addWidget(info);
 }
 void page_notes::create_dock(){
     dock=new QWidget;
@@ -23,55 +37,4 @@ void page_notes::create_dock(){
     layout->addStretch();
     layout->addWidget(lecture_seul);
     layout->addWidget(editer);
-}
-page_article::page_article(Article &n): page_notes(n){
-    description=new QTextEdit;
-    description->setEnabled(false);
-    description->setDocument(&(n.getText()));
-    layout_p->addStretch();
-    layout_p->addWidget(description);
-    addInfo();
-
-}
-page_Task::page_Task(Task &n): page_notes(n){
-    action=new QLabel ("Action");
-    E_action=new QLineEdit;
-
-    duedate=new QGroupBox ();
-    duedate->setCheckable(true);
-    duedate->setTitle("Date de réalisation :");
-    optional_duedate=new QHBoxLayout();
-    duedate->setLayout(optional_duedate);
-    E_duedate=new QDateTimeEdit();
-    E_duedate->setDate(QDate::currentDate());
-    E_duedate->setTime(QTime::currentTime());
-    E_duedate->setCalendarPopup(true);
-    optional_duedate->addStretch();
-    optional_duedate->addWidget(E_duedate);
-    priority=new QGroupBox ();
-    priority->setCheckable(true);
-    priority->setTitle("priorité :");
-    optional_priority=new QHBoxLayout;
-    priority->setLayout(optional_priority);
-    E_priority=new QSpinBox();
-    optional_priority->addStretch();
-    optional_priority->addWidget(E_priority);
-
-    status=new QLabel("Etat");
-    E_status=new QComboBox;
-    E_status->addItem("En cours");
-    E_status->addItem("En attente");
-    E_status->addItem("Complétée");
-
-    grid=new QGridLayout(this);
-    grid->addWidget(action,0,0);
-    grid->addWidget(E_action,0,1);
-    grid->addWidget(status,1,0);
-    grid->addWidget(E_status,1,1);
-    grid->addWidget(duedate,2,0,2,2);
-    grid->addWidget(priority,3,0,3,2);
-
-    layout_p->addLayout(grid);
-    addInfo();
-
 }
